pyproject.toml
[project]
name = "github-topics-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.6.0",
]

[tool.ruff]
# Target Python 3.13+
target-version = "py313"

# Include source files
include = ["src/**/*.py", "scripts/**/*.py"]

# Exclude common directories
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
]

# Line length
line-length = 100

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Disable specific rules
ignore = [
    "E501",  # Line too long (handled by formatter)
    "W191",  # Indentation contains tabs (we prefer tabs)
]

[tool.ruff.lint.isort]
# Sort imports
known-first-party = ["src"]

[tool.ruff.format]
# Use double quotes
quote-style = "double"

# Indent with tabs
indent-style = "tab"
